+++ core/components/romanesco/docs/changelog.md
@@ -4,6 +4,7 @@
 Released on ...
 
 Fixes and improvements:
+- Add system default setting for CB layout backgrounds
 - Add non-white class to body if custom background is set
 
 ## Romanesco Patterns 0.15.1